Deleted setup.go and refactored service files.
Removed the setup.go for database configuration. Renamed service files for consistency and updated variable names for better clarity. Corrected datatype mismatches in models and service logic, improving overall code maintainability.

+++ models/product.go
@@ -12,6 +12,6 @@ type Product struct {
 	Ncm          int     `json:"ncm"`
 	Category     string  `gorm:"size:50;not null; unique" json:"category"`
 	Inventory    int     `gorm:"not null;" json:"inventory"`
-	Price        float64 `gorm:"not null;" json:"price"`
-	Cost         float64 `gorm:"not null;" json:"cost"`
+	Price        float32 `gorm:"not null;" json:"price"`
+	Cost         float32 `gorm:"not null;" json:"cost"`
 }
